Use Browser HTTP Client Adapter for Web
